+++ libs/cardano-ledger-binary/src/Cardano/Ledger/Binary/Decoding/DecCBOR.hs
@@ -1,6 +1,5 @@
 {-# LANGUAGE BangPatterns #-}
 {-# LANGUAGE CPP #-}
-{-# LANGUAGE DefaultSignatures #-}
 {-# LANGUAGE DerivingStrategies #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
@@ -19,7 +18,7 @@ module Cardano.Ledger.Binary.Decoding.DecCBOR (
   decodeScriptContextFromData,
 ) where
 
-import qualified Cardano.Binary as Plain (Decoder, FromCBOR (..))
+import qualified Cardano.Binary as Plain (Decoder)
 import Cardano.Crypto.DSIGN.Class (
   DSIGNAlgorithm,
   SigDSIGN,
@@ -27,7 +26,7 @@ import Cardano.Crypto.DSIGN.Class (
   SignedDSIGN,
   VerKeyDSIGN,
  )
-import Cardano.Crypto.Hash.Class (Hash, HashAlgorithm)
+import Cardano.Crypto.Hash.Class (Hash, HashAlgorithm, hashFromBytesShort)
 import Cardano.Crypto.KES.Class (KESAlgorithm, SigKES, VerKeyKES)
 import Cardano.Crypto.VRF.Class (
   CertVRF,
@@ -92,9 +91,6 @@ import Prelude hiding (decodeFloat)
 
 class Typeable a => DecCBOR a where
   decCBOR :: Decoder s a
-  default decCBOR :: Plain.FromCBOR a => Decoder s a
-  decCBOR = fromPlainDecoder Plain.fromCBOR
-  {-# INLINE decCBOR #-}
 
   -- | Validate decoding of a Haskell value, without the need to actually construct
   -- it. Could be slightly faster than `decCBOR`, however it should respect this law:
@@ -476,7 +472,11 @@ instance (DSIGNAlgorithm v, Typeable a) => DecCBOR (SignedDSIGN v a) where
 -- Hash
 --------------------------------------------------------------------------------
 
-instance (HashAlgorithm h, Typeable a) => DecCBOR (Hash h a)
+instance (HashAlgorithm h, Typeable a) => DecCBOR (Hash h a) where
+  decCBOR = do
+    sbs <- decCBOR @ShortByteString
+    maybe (fail "Invalid hash size") pure $ hashFromBytesShort sbs
+  {-# INLINE decCBOR #-}
 
 --------------------------------------------------------------------------------
 -- KES
@@ -530,13 +530,21 @@ instance DecCBOR Praos.VerKey where
   decCBOR = decCBOR >>= Praos.vkFromBytes
   {-# INLINE decCBOR #-}
 
-deriving instance DecCBOR (VerKeyVRF Praos.PraosVRF)
+instance DecCBOR (VerKeyVRF Praos.PraosVRF) where
+  decCBOR = decodeVerKeyVRF
+  {-# INLINE decCBOR #-}
 
-deriving instance DecCBOR (SignKeyVRF Praos.PraosVRF)
+instance DecCBOR (SignKeyVRF Praos.PraosVRF) where
+  decCBOR = decodeSignKeyVRF
+  {-# INLINE decCBOR #-}
 
-deriving instance DecCBOR (CertVRF Praos.PraosVRF)
+instance DecCBOR (CertVRF Praos.PraosVRF) where
+  decCBOR = decodeCertVRF
+  {-# INLINE decCBOR #-}
 
-deriving instance Typeable v => DecCBOR (OutputVRF v)
+instance Typeable v => DecCBOR (OutputVRF v) where
+  decCBOR = OutputVRF <$> decCBOR
+  {-# INLINE decCBOR #-}
 
 instance (VRFAlgorithm v, Typeable a) => DecCBOR (CertifiedVRF v a) where
   decCBOR =