+++ libs/cardano-ledger-core/src/Cardano/Ledger/Coin.hs
@@ -57,6 +57,7 @@ import Cardano.Ledger.Binary (
   FromCBOR (..),
   ToCBOR,
   decodeWord64,
+  fromPlainDecoder,
  )
 import qualified Cardano.Ledger.Binary.Plain as Plain
 import Cardano.Ledger.Compactible
@@ -98,7 +99,9 @@ newtype Coin = Coin {unCoin :: Integer}
 instance FromCBOR Coin where
   fromCBOR = Coin . toInteger <$> Plain.decodeWord64
 
-instance DecCBOR Coin
+instance DecCBOR Coin where
+  decCBOR = fromPlainDecoder fromCBOR
+  {-# INLINE decCBOR #-}
 
 newtype DeltaCoin = DeltaCoin Integer
   deriving (Eq, Ord, Generic, Enum, NoThunks)